Add option to enable/disable background polling on a sensor instance.
This greatly reduces overhead if the change events are not used.

A possible future enhancement would be to have the sensor instance monitor for addListener/removeListener events and only enable polling when a listener is present, however this would technically break API compatibility by causing the 'lastTemp' property to stop updating if a change listener is not present.